+++ services/invoices.service.ts
@@ -60,13 +60,20 @@ const TheService: ServiceSchema = {
         if (instance.errors) {
           throw new MoleculerError('Store not found', 404);
         }
+
+        const invoiceParams: {[key:string]: string} = {
+          customerId: instance.internal_data.omsId,
+          discount: ctx.params.discount && ctx.params.discount.value,
+          discountType: ctx.params.discount && ctx.params.discount.type,
+          items: ctx.params.items,
+        };
+
+        if (ctx.params.coupon) {
+          invoiceParams.coupon = ctx.params.coupon;
+        }
+
         return ctx
-          .call('oms.createInvoice', {
-            customerId: instance.internal_data.omsId,
-            discount: ctx.params.discount && ctx.params.discount.value,
-            discountType: ctx.params.discount && ctx.params.discount.type,
-            items: ctx.params.items,
-          })
+          .call('oms.createInvoice', invoiceParams)
           .then(
             res => {
               this.broker.cacher.clean(`invoices.get:${instance.consumer_key}*`);
@@ -81,9 +88,14 @@ const TheService: ServiceSchema = {
     applyCredits: {
       auth: 'Bearer',
       async handler(ctx: Context) {
-        const instance = await ctx.call('stores.findInstance', {
-          consumerKey: ctx.meta.user,
-        });
+        const instance = await ctx.call('stores.me');
+        const {params} = ctx;
+        if(params.useSavedPaymentMethods && instance.credit < params.paymentAmount) {
+          await ctx.call('paymentGateway.charge', {
+            storeId: instance.url,
+            amount: params.paymentAmount - instance.credit,
+          });
+        }
         if (instance.errors) {
           throw new MoleculerError('Store not found', 404);
         }
@@ -166,6 +178,7 @@ const TheService: ServiceSchema = {
         last_modified_time: invoice.lastModifiedTime,
         shipping_charge: invoice.shippingCharge,
         adjustment: invoice.adjustment,
+        coupon: invoice.coupon,
       };
     },
   },