+++ services/subscription-cron.service.ts
@@ -1,5 +1,6 @@
 import { Context, ServiceSchema } from 'moleculer';
 import * as Cron from 'moleculer-cron';
+import { Store } from '../utilities/types';
 
 const TheService: ServiceSchema = {
   name: 'subscription-cron',
@@ -36,6 +37,7 @@ const TheService: ServiceSchema = {
         if (!subscription) {
           return null;
         }
+
         try {
           const createSubResponse = await ctx.call('subscription.create', {
             storeId: subscription.storeId,