+++ utilities/mixins/openapi/membership.openapi.mixin.ts
@@ -245,6 +245,14 @@ const MembershipGetOpenapi = {
         type: 'string',
       },
     },
+    {
+      name: 'country',
+      'in': 'query',
+      required: false,
+      schema: {
+        type: 'string',
+      },
+    },
   ],
   responses: {
     200: {