+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## 1.5.1
+
+### 🚀 Features
+
+- Payment end-points
+- Apply credits by saved payments methods
+- Automatically charge saved cards while adding subscription
+- Subscription charge try
+- Auto renew subscription query
+- Get product by sku with currency option
+- Store logo
+- Send store logo with create order
+- Add "subscription, storeStatus and lastOrderDate" to update crm
+- Add country option to Memberships and calculate the taxes according to the country
+- Send order warnings to Zoho & knawat support
+
+### 🐛 Fixes
+
+- Coupon list docs & validation
+- Order items limit
+- Openapi docs applyCredits to bearer auth
+- Add total & tax total fields to order response
+- Send coupon with subscription invoice
+- List subscription pagination
+
+### ✨ Enhancements
+
+- Products example doc
+- Nats not used from docker
+- Get products by variation limit
+- Use getTaxWithCalc method
+- Add uninstalled option to store status
 
 ## 1.4.1